+++ src/include/parser/expression/case_expression.h
@@ -5,7 +5,6 @@
 #include <vector>
 #include "common/macros.h"
 #include "parser/expression/abstract_expression.h"
-#include "parser/expression_defs.h"
 
 namespace terrier::parser {
