+++ src/include/parser/expression/star_expression.h
@@ -3,8 +3,6 @@
 #include <memory>
 #include <vector>
 #include "parser/expression/abstract_expression.h"
-#include "parser/expression_defs.h"
-#include "type/type_id.h"
 
 namespace terrier::parser {
 /**
