+++ src/include/type/type_util.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include "common/exception.h"
 #include "common/strong_typedef.h"
 #include "storage/block_layout.h"
 #include "type/type_id.h"
